chore: removed unused code
