--- Gopkg.toml
@@ -1,4 +1,0 @@
-
-[[constraint]]
-  name = "github.com/volatiletech/authboss"
-  version = "2.0.0"
