+++ api/user.go
@@ -1,15 +1,7 @@
 package api
 
 import (
-	"errors"
 	"net/http"
-
-	"github.com/go-chi/jwtauth"
-
-	"github.com/asdine/storm/q"
-	"github.com/go-chi/render"
-
-	"github.com/smarthut/smarthut/model"
 )
 
 // func (api *API) userCtx(next http.Handler) http.Handler {
@@ -46,34 +38,4 @@ func (api *API) updateUser(w http.ResponseWriter, r *http.Request) {
 
 func (api *API) deleteUser(w http.ResponseWriter, r *http.Request) {
 
-}
-
-func (api *API) authenticate(w http.ResponseWriter, r *http.Request) {
-	var cred model.Credentials
-	if err := render.DecodeJSON(r.Body, &cred); err != nil {
-		handleError(err, w, r)
-		return
-	}
-
-	// Try to find with Username or Email provided in the Login field
-	var user model.User
-	if err := api.db.Select(q.Or(q.StrictEq("Username", cred.Login), q.StrictEq("Email", cred.Login))).First(&user); err != nil {
-		handleError(err, w, r)
-		return
-	}
-
-	// Validate password provided
-	if !user.Authenticate(cred.Password) {
-		handleError(errors.New("password is not valid"), w, r)
-		return
-	}
-
-	// Generate a new JWT token
-	_, tokenString, err := api.tokenAuth.Encode(jwtauth.Claims{"user_id": user.Username})
-	if err != nil {
-		handleError(err, w, r)
-		return
-	}
-
-	w.Write([]byte(tokenString))
 }
